create agent editor group from group resource
